% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_iris.R
\name{predict_iris}
\alias{predict_iris}
\title{Iris classification on trained xgboost model}
\usage{
predict_iris(Sepal.Length = 0, Sepal.Width = 0, Petal.Length = 0,
  Petal.Width = 0)
}
\description{
Iris classification on trained xgboost model
}
